+++ lib/ruby_jard/repl_proxy.rb
@@ -1,15 +1,6 @@
 # frozen_string_literal: true
 
-require 'ruby_jard/commands/validation_helpers'
-require 'ruby_jard/commands/continue_command'
-require 'ruby_jard/commands/up_command'
-require 'ruby_jard/commands/down_command'
-require 'ruby_jard/commands/next_command'
-require 'ruby_jard/commands/step_command'
-require 'ruby_jard/commands/step_out_command'
-require 'ruby_jard/commands/frame_command'
-require 'ruby_jard/commands/list_command'
-require 'ruby_jard/commands/color_scheme_command'
+require 'pty'
 
 module RubyJard
   ##
@@ -45,7 +36,6 @@ module RubyJard
       'stat',          # Included in jard UI
       'backtrace',     # Re-implemented later
       'break',         # Re-implemented later
-      'exit',          # Conflicted with continue
       'exit-all',      # Conflicted with continue
       'exit-program',  # We already have `exit` native command
       '!pry',          # No need to complicate things
@@ -58,7 +48,7 @@ module RubyJard
     COMMANDS = [
       CMD_FLOW      = :flow,
       CMD_EVALUATE  = :evaluate,
-      CMD_IDLE      = :idle,
+      CMD_EVALUATED = :evaluated,
       CMD_INTERRUPT = :interrupt
     ].freeze
     INTERNAL_KEY_BINDINGS = {
@@ -66,53 +56,121 @@ module RubyJard
       RubyJard::Keys::CTRL_C   => (KEY_BINDING_INTERRUPT = :interrupt)
     }.freeze
 
-    KEYPRESS_POLLING = 0.1 # 100ms
+    KEY_READ_TIMEOUT = 0.1           # 100ms
+    PTY_OUTPUT_TIMEOUT = 1.to_f / 60 # 60hz
+
+    STATES = [
+      STATE_READY = 0,
+      STATE_EXITING = 1,
+      STATE_EXITED = 2
+    ].freeze
 
     def initialize(key_bindings: nil)
-      @pry_read_stream, @pry_write_stream = IO.pipe
+      @state = STATE_EXITED
+
+      @pry_input_pipe_read, @pry_input_pipe_write = IO.pipe
+      @pry_output_pty_read, @pry_output_pty_write = PTY.open
       @pry = pry_instance
-      @commands = Queue.new
+      @pry_commands = Queue.new
+
       @key_bindings = key_bindings || RubyJard::KeyBindings.new
       INTERNAL_KEY_BINDINGS.each do |sequence, action|
         @key_bindings.push(sequence, action)
       end
-    end
 
-    def read_key
-      RubyJard::Console.getch(STDIN, KEYPRESS_POLLING)
+      @pry_pty_output_thread = Thread.new { pry_pty_output }
     end
 
     def repl(current_binding)
-      Readline.input = @pry_read_stream
-      @commands.clear
+      ready!
+
+      RubyJard::Console.disable_echo!
+
+      Readline.input = @pry_input_pipe_read
+      Readline.output = @pry_output_pty_write
+      @pry_commands.clear
       @pry.binding_stack.clear
 
-      pry_thread = Thread.new do
-        pry_repl(current_binding)
-      end
-      pry_thread.report_on_exception = false if pry_thread.respond_to?(:report_on_exception)
+      @pry_input_thread = Thread.new { pry_repl(current_binding) }
+      @pry_input_thread.report_on_exception = false if @pry_input_thread.respond_to?(:report_on_exception)
       loop do
-        break unless pry_thread.alive?
+        break unless ready?
 
-        if @commands.empty?
+        if @pry_commands.empty?
           listen_key_press
         else
-          cmd, value = @commands.deq
-          handle_command(pry_thread, cmd, value)
+          cmd, value = @pry_commands.deq
+          handle_command(cmd, value)
         end
       end
-      pry_thread&.join
+    ensure
+      RubyJard::Console.enable_echo!
       Readline.input = STDIN
+      Readline.output = STDOUT
+      @pry_input_thread&.exit if @pry_input_thread&.alive?
+
+      exited!
+    end
+
+    private
+
+    def ready?
+      @state == STATE_READY
+    end
+
+    def ready!
+      @state = STATE_READY
+    end
+
+    def exiting?
+      @state == STATE_EXITING
+    end
+
+    def exiting!
+      @state = STATE_EXITING
+    end
+
+    def exited?
+      @state == STATE_EXITED
+    end
+
+    def exited!
+      @state = STATE_EXITED
+    end
+
+    def read_key
+      RubyJard::Console.getch(STDIN, KEY_READ_TIMEOUT)
     end
 
     def pry_repl(current_binding)
+      sleep PTY_OUTPUT_TIMEOUT unless ready?
+
       flow = RubyJard::ControlFlow.listen do
         @pry.repl(current_binding)
       end
-      @commands << [CMD_FLOW, flow]
+      @pry_commands << [CMD_FLOW, flow]
     rescue StandardError => e
       RubyJard::ScreenManager.draw_error(e)
       raise
+    end
+
+    def pry_pty_output
+      loop do
+        if exiting?
+          if @pry_output_pty_read.ready?
+            STDOUT.print @pry_output_pty_read.read_nonblock(255)
+          else
+            exited!
+          end
+        elsif exited?
+          sleep PTY_OUTPUT_TIMEOUT
+        else
+          STDOUT.print @pry_output_pty_read.read_nonblock(255)
+        end
+      rescue IO::WaitReadable, IO::WaitWritable
+        # Retry
+        sleep PTY_OUTPUT_TIMEOUT
+      end
     end
 
     def listen_key_press
@@ -120,52 +178,54 @@ module RubyJard
       if key.is_a?(RubyJard::KeyBinding)
         handle_key_binding(key)
       elsif !key.empty?
-        @pry_write_stream.write(key)
+        @pry_input_pipe_write.write(key)
       end
     end
 
     def handle_key_binding(key_binding)
       case key_binding.action
       when KEY_BINDING_ENDLINE
-        @pry_write_stream.write(key_binding.sequence)
-        @commands << [CMD_EVALUATE]
+        @pry_input_pipe_write.write(key_binding.sequence)
+        @pry_commands << [CMD_EVALUATE]
       when KEY_BINDING_INTERRUPT
-        @commands << [CMD_INTERRUPT]
+        handle_interrupt_command
       else
-        @commands << [
-          CMD_FLOW, RubyJard::ControlFlow.new(:key_binding, action: key_binding.action)
-        ]
+        handle_flow_command(RubyJard::ControlFlow.new(:key_binding, action: key_binding.action))
       end
     end
 
-    def handle_command(pry_thread, cmd, value)
+    def handle_command(cmd, value)
       case cmd
       when CMD_FLOW
-        pry_thread.exit if pry_thread.alive?
-        RubyJard::ControlFlow.dispatch(value)
+        handle_flow_command(value)
       when CMD_EVALUATE
         loop do
-          cmd, value = @commands.deq
-          break if [CMD_IDLE, CMD_FLOW, CMD_INTERRUPT].include?(cmd)
+          cmd, value = @pry_commands.deq
+          break if [CMD_EVALUATED, CMD_FLOW].include?(cmd)
         end
-        handle_command(pry_thread, cmd, value)
-      when CMD_INTERRUPT
-        handle_interrupt_command(pry_thread)
-      when CMD_IDLE
+        handle_command(cmd, value)
+      when CMD_EVALUATED
         # Ignore
       end
     end
 
-    def handle_interrupt_command(pry_thread)
-      pry_thread.raise Interrupt if pry_thread.alive?
+    def handle_interrupt_command
+      @pry_input_thread&.raise Interrupt if @pry_input_thread&.alive?
       loop do
         begin
-          sleep 0.1
+          sleep PTY_OUTPUT_TIMEOUT
         rescue Interrupt
           # Interrupt spam. Ignore.
         end
-        break unless pry_thread.pending_interrupt?
+        break unless @pry_input_thread&.pending_interrupt?
       end
+    end
+
+    def handle_flow_command(flow)
+      @pry_commands.clear
+      exiting!
+      sleep PTY_OUTPUT_TIMEOUT until exited?
+      RubyJard::ControlFlow.dispatch(flow)
     end
 
     def pry_instance
@@ -173,7 +233,8 @@ module RubyJard
         prompt: pry_jard_prompt,
         quiet: true,
         commands: pry_command_set,
-        hooks: pry_hooks
+        hooks: pry_hooks,
+        output: @pry_output_pty_write
       )
       # I'll be burned in hell for this
       # TODO: Contact pry author to add :after_handle_line hook
@@ -214,17 +275,15 @@ module RubyJard
     def pry_hooks
       hooks = Pry::Hooks.default
       hooks.add_hook(:after_read, :jard_proxy_acquire_lock) do |read_string, _pry|
-        @commands <<
-          if Pry::Code.complete_expression?(read_string)
-            [CMD_EVALUATE]
-          else
-            [CMD_IDLE]
-          end
+        unless Pry::Code.complete_expression?(read_string)
+          @pry_commands << [CMD_EVALUATED]
+        end
       rescue SyntaxError
-        @commands << [CMD_IDLE]
+        # Ignore
+        @pry_commands << [CMD_EVALUATED]
       end
       hooks.add_hook(:after_handle_line, :jard_proxy_release_lock) do
-        @commands << [CMD_IDLE]
+        @pry_commands << [CMD_EVALUATED]
       end
     end
   end