+++ spec/ruby_jard/screens/variables/variables_screen_spec.rb
@@ -3,7 +3,7 @@
 RSpec.describe 'RubyJard::Screens::VariablesScreen' do
   let(:work_dir) { File.join(RSPEC_ROOT, '/ruby_jard/screens/variables') }
 
-  context 'with TOPLEVEL_BINDING variables' do
+  context 'when jard stops at top-level binding' do
     let(:expected_output_1) do
       <<~EXPECTED
         ┌ Variables ───────────────────────────────────────────────────────────────────┐
@@ -16,6 +16,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
         │  variable_f = 1.1                                                            │
         │  variable_g = 99..100                                                        │
         │• variable_h = nil                                                            │
+        │  variable_k = #<StandardError: A random error>                               │
         └──────────────────────────────────────────────────────────────────────────────┘
       EXPECTED
     end
@@ -32,6 +33,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
         │  variable_f = 1.1                                                            │
         │  variable_g = 99..100                                                        │
         │  variable_h = 15                                                             │
+        │  variable_k = #<StandardError: A random error>                               │
         └──────────────────────────────────────────────────────────────────────────────┘
       EXPECTED
     end
@@ -48,6 +50,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
         │• variable_f = 1.1                                                            │
         │  variable_g = 99..100                                                        │
         │• variable_h = 15                                                             │
+        │  variable_k = #<StandardError: A random error>                               │
         └──────────────────────────────────────────────────────────────────────────────┘
       EXPECTED
     end
@@ -65,7 +68,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
     end
   end
 
-  context 'with context inside an instance method' do
+  context 'when jard stops at an instance method' do
     let(:expected_output_1) do
       <<~EXPECTED
         ┌ Variables ───────────────────────────────────────────────────────────────────┐
@@ -124,7 +127,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
     end
   end
 
-  context 'with context inside a class method' do
+  context 'when jard stops inside a class method' do
     let(:expected_output) do
       <<~EXPECTED
         ┌ Variables ───────────────────────────────────────────────────────────────────┐
@@ -146,7 +149,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
     end
   end
 
-  context 'with context inside a nested loop' do
+  context 'when jard stops inside a nested loop' do
     let(:expected_output) do
       <<~EXPECTED
         ┌ Variables ───────────────────────────────────────────────────────────────────┐
@@ -235,7 +238,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
     end
   end
 
-  context 'with code evaluation' do
+  context 'when jard steps into a code evaluation' do
     let(:expected_output_1) do
       <<~'EXPECTED'
         ┌ Variables ───────────────────────────────────────────────────────────────────┐
@@ -280,7 +283,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
     end
   end
 
-  context 'when stop at the end of a method' do
+  context 'when jard stops at the end of a method' do
     let(:expected_output_1) do
       <<~EXPECTED
         ┌ Variables ───────────────────────────────────────────────────────────────────┐
@@ -311,7 +314,7 @@ RSpec.describe 'RubyJard::Screens::VariablesScreen' do
     end
   end
 
-  context 'when run with ruby -e' do
+  context 'when use jard with ruby -e' do
     let(:expected_output_1) do
       <<~EXPECTED
         ┌ Variables ───────────────────────────────────────────────────────────────────┐